+++ components-ui/app/src/tol-ui/src/hooks/useBoard.ts
@@ -19,6 +19,8 @@ export const useBoard = () => {
       setPrivilege: () => { },
       editMode: false,
       setEditMode: () => { },
+      tableLoading: false,
+      setTableLoading: () => {},
       layoutMode: false,
       setLayoutMode: () => { }
     };